Add fade animations, Move blink animation to `Functions.kt`
