+++ app/src/main/java/com/finnmglas/launcher/FirstStartupActivity.kt
@@ -10,22 +10,6 @@ import android.view.animation.Animation
 import androidx.appcompat.app.AppCompatActivity
 import kotlinx.android.synthetic.main.activity_firststartup.*
 
-// Taken from https://stackoverflow.com/questions/47293269
-fun View.blink(
-    times: Int = Animation.INFINITE,
-    duration: Long = 1000L,
-    offset: Long = 20L,
-    minAlpha: Float = 0.2f,
-    maxAlpha: Float = 1.0f,
-    repeatMode: Int = Animation.REVERSE
-) {
-    startAnimation(AlphaAnimation(minAlpha, maxAlpha).also {
-        it.duration = duration
-        it.startOffset = offset
-        it.repeatMode = repeatMode
-        it.repeatCount = times
-    })
-}
 
 class FirstStartupActivity : AppCompatActivity(){
